Refactor few widgets to dedicated modules
FilteredList, TabBar, Tab, RangeSlider, DimensionsTooltip
