+++ library/lua/gui/widgets/tab_bar.lua
@@ -1,0 +1,210 @@
+local Widget = require('gui.widgets.widget')
+local ResizingPanel = require('gui.widgets.resizing_panel')
+
+local to_pen = dfhack.pen.parse
+
+---@class widgets.TabPens
+---@field text_mode_tab_pen dfhack.pen
+---@field text_mode_label_pen dfhack.pen
+---@field lt dfhack.pen
+---@field lt2 dfhack.pen
+---@field t dfhack.pen
+---@field rt2 dfhack.pen
+---@field rt dfhack.pen
+---@field lb dfhack.pen
+---@field lb2 dfhack.pen
+---@field b dfhack.pen
+---@field rb2 dfhack.pen
+---@field rb dfhack.pen
+
+local TSO = df.global.init.tabs_texpos[0] -- tab spritesheet offset
+local DEFAULT_ACTIVE_TAB_PENS = {
+    text_mode_tab_pen=to_pen{fg=COLOR_YELLOW},
+    text_mode_label_pen=to_pen{fg=COLOR_WHITE},
+    lt=to_pen{tile=TSO+5, write_to_lower=true},
+    lt2=to_pen{tile=TSO+6, write_to_lower=true},
+    t=to_pen{tile=TSO+7, fg=COLOR_BLACK, write_to_lower=true, top_of_text=true},
+    rt2=to_pen{tile=TSO+8, write_to_lower=true},
+    rt=to_pen{tile=TSO+9, write_to_lower=true},
+    lb=to_pen{tile=TSO+15, write_to_lower=true},
+    lb2=to_pen{tile=TSO+16, write_to_lower=true},
+    b=to_pen{tile=TSO+17, fg=COLOR_BLACK, write_to_lower=true, bottom_of_text=true},
+    rb2=to_pen{tile=TSO+18, write_to_lower=true},
+    rb=to_pen{tile=TSO+19, write_to_lower=true},
+}
+
+local DEFAULT_INACTIVE_TAB_PENS = {
+    text_mode_tab_pen=to_pen{fg=COLOR_BROWN},
+    text_mode_label_pen=to_pen{fg=COLOR_DARKGREY},
+    lt=to_pen{tile=TSO+0, write_to_lower=true},
+    lt2=to_pen{tile=TSO+1, write_to_lower=true},
+    t=to_pen{tile=TSO+2, fg=COLOR_WHITE, write_to_lower=true, top_of_text=true},
+    rt2=to_pen{tile=TSO+3, write_to_lower=true},
+    rt=to_pen{tile=TSO+4, write_to_lower=true},
+    lb=to_pen{tile=TSO+10, write_to_lower=true},
+    lb2=to_pen{tile=TSO+11, write_to_lower=true},
+    b=to_pen{tile=TSO+12, fg=COLOR_WHITE, write_to_lower=true, bottom_of_text=true},
+    rb2=to_pen{tile=TSO+13, write_to_lower=true},
+    rb=to_pen{tile=TSO+14, write_to_lower=true},
+}
+
+---------
+-- Tab --
+---------
+
+---@class widgets.Tab.attrs: widgets.Widget.attrs
+---@field id? string|integer
+---@field label string
+---@field on_select? function
+---@field get_pens? fun(): widgets.TabPens
+
+---@class widgets.Tab.attrs.partial: widgets.Tab.attrs
+
+---@class widgets.Tab.initTable: widgets.Tab.attrs
+---@field label string
+
+---@class widgets.Tab: widgets.Widget, widgets.Tab.attrs
+---@field super widgets.Widget
+---@field ATTRS widgets.Tab.attrs|fun(attributes: widgets.Tab.attrs.partial)
+---@overload fun(init_table: widgets.Tab.initTable): self
+Tab = defclass(Tabs, Widget)
+Tab.ATTRS{
+    id=DEFAULT_NIL,
+    label=DEFAULT_NIL,
+    on_select=DEFAULT_NIL,
+    get_pens=DEFAULT_NIL,
+}
+
+function Tab:preinit(init_table)
+    init_table.frame = init_table.frame or {}
+    init_table.frame.w = #init_table.label + 4
+    init_table.frame.h = 2
+end
+
+function Tab:onRenderBody(dc)
+    local pens = self.get_pens()
+    dc:seek(0, 0)
+    if dfhack.screen.inGraphicsMode() then
+        dc:char(nil, pens.lt):char(nil, pens.lt2)
+        for i=1,#self.label do
+            dc:char(self.label:sub(i,i), pens.t)
+        end
+        dc:char(nil, pens.rt2):char(nil, pens.rt)
+        dc:seek(0, 1)
+        dc:char(nil, pens.lb):char(nil, pens.lb2)
+        for i=1,#self.label do
+            dc:char(self.label:sub(i,i), pens.b)
+        end
+        dc:char(nil, pens.rb2):char(nil, pens.rb)
+    else
+        local tp = pens.text_mode_tab_pen
+        dc:char(' ', tp):char('/', tp)
+        for i=1,#self.label do
+            dc:char('-', tp)
+        end
+        dc:char('\\', tp):char(' ', tp)
+        dc:seek(0, 1)
+        dc:char('/', tp):char('-', tp)
+        dc:string(self.label, pens.text_mode_label_pen)
+        dc:char('-', tp):char('\\', tp)
+    end
+end
+
+function Tab:onInput(keys)
+    if Tab.super.onInput(self, keys) then return true end
+    if keys._MOUSE_L and self:getMousePos() then
+        self.on_select(self.id)
+        return true
+    end
+end
+
+-------------
+-- Tab Bar --
+-------------
+
+---@class widgets.TabBar.attrs: widgets.ResizingPanel.attrs
+---@field labels string[]
+---@field on_select? function
+---@field get_cur_page? function
+---@field active_tab_pens widgets.TabPens
+---@field inactive_tab_pens widgets.TabPens
+---@field get_pens? fun(index: integer, tabbar: self): widgets.TabPens
+---@field key string
+---@field key_back string
+
+---@class widgets.TabBar.attrs.partial: widgets.TabBar.attrs
+
+---@class widgets.TabBar.initTable: widgets.TabBar.attrs
+---@field labels string[]
+
+---@class widgets.TabBar: widgets.ResizingPanel, widgets.TabBar.attrs
+---@field super widgets.ResizingPanel
+---@field ATTRS widgets.TabBar.attrs|fun(attribute: widgets.TabBar.attrs.partial)
+---@overload fun(init_table: widgets.TabBar.initTable): self
+TabBar = defclass(TabBar, ResizingPanel)
+TabBar.ATTRS{
+    labels=DEFAULT_NIL,
+    on_select=DEFAULT_NIL,
+    get_cur_page=DEFAULT_NIL,
+    active_tab_pens=DEFAULT_ACTIVE_TAB_PENS,
+    inactive_tab_pens=DEFAULT_INACTIVE_TAB_PENS,
+    get_pens=DEFAULT_NIL,
+    key='CUSTOM_CTRL_T',
+    key_back='CUSTOM_CTRL_Y',
+}
+
+---@param self widgets.TabBar
+function TabBar:init()
+    for idx,label in ipairs(self.labels) do
+        self:addviews{
+            Tab{
+                frame={t=0, l=0},
+                id=idx,
+                label=label,
+                on_select=self.on_select,
+                get_pens=self.get_pens and function()
+                    return self.get_pens(idx, self)
+                end or function()
+                    if self.get_cur_page() == idx then
+                        return self.active_tab_pens
+                    end
+
+                    return self.inactive_tab_pens
+                end,
+            }
+        }
+    end
+end
+
+function TabBar:postComputeFrame(body)
+    local t, l, width = 0, 0, body.width
+    for _,tab in ipairs(self.subviews) do
+        if l > 0 and l + tab.frame.w > width then
+            t = t + 2
+            l = 0
+        end
+        tab.frame.t = t
+        tab.frame.l = l
+        l = l + tab.frame.w
+    end
+end
+
+function TabBar:onInput(keys)
+    if TabBar.super.onInput(self, keys) then return true end
+    if self.key and keys[self.key] then
+        local zero_idx = self.get_cur_page() - 1
+        local next_zero_idx = (zero_idx + 1) % #self.labels
+        self.on_select(next_zero_idx + 1)
+        return true
+    end
+    if self.key_back and keys[self.key_back] then
+        local zero_idx = self.get_cur_page() - 1
+        local prev_zero_idx = (zero_idx - 1) % #self.labels
+        self.on_select(prev_zero_idx + 1)
+        return true
+    end
+end
+
+TabBar.Tab = Tab
+
+return TabBar
